fix: review CRUD 수정 (#190)
* feat: review, store CRUD 추가

* feat: review, store CRUD 추가

* feat: Local entity 수정

* feat: Local entity 수정

* fix: review CRUD 수정

+++ src/main/java/com/main_39/Spring/review/dto/ReviewsResponseDto.java
@@ -12,11 +12,11 @@ public class ReviewsResponseDto {
     private List<ReviewStoreResponseDto> reviews;
     private List<CommentResponseDto> comment;
     private int totalReview;
-    private int totalGrade;
+    private double totalGrade;
     private int totalComment;
 
     @Builder
-    public ReviewsResponseDto(Long storeId, int totalReview, int totalGrade, int totalComment,
+    public ReviewsResponseDto(Long storeId, int totalReview, double totalGrade, int totalComment,
                               List<ReviewStoreResponseDto> reviews, List<CommentResponseDto> comment) {
         this.storeId = storeId;
         this.reviews = reviews;
